refactor: remove dead code and useless comments
